Refactor ingress helpers
