--- ingress/paths.py
@@ -1,7 +1,0 @@
-import os
-
-BASE_CL_DIR = os.path.join("data", "cl")
-
-CLUSTER_PATH = "clusters"
-OPINION_PATH = "opinions"
-CITATIONS_PATH = "citations.csv"
